Shop and Item model and multer cloudinary issue fixed

+++ server/src/utils/cloudinary.js
@@ -1,5 +1,6 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
+import { ApiError } from "./api-error.js";
 
 cloudinary.config({
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
@@ -7,15 +8,28 @@ cloudinary.config({
   api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
-const uploadToCloudinary = async (file) => {
+export const uploadToCloudinary = async (localFilePath, folder = "uploads") => {
   try {
-    const result = await cloudinary.uploader.upload(file);
+    if (!localFilePath) return null;
+
+    const result = await cloudinary.uploader.upload(localFilePath, {
+      folder: folder,
+    });
+
     fs.unlinkSync(file);
-    return result.secure_url;
+
+    return result;
   } catch (error) {
     fs.unlinkSync(file);
     console.log("Error in uploading to cloudinary", error);
   }
 };
 
-export default uploadToCloudinary;
+export const deleteFromCloudinary = async (publicId) => {
+  try {
+    await cloudinary.uploader.destroy(publicId);
+  } catch (error) {
+    console.log("Error in deleting from cloudinary", error);
+    throw new ApiError(500, "Error in deleting from cloudinary");
+  }
+};
